fixed bug in seqtk's trimfq that didn't trim entire runs of low qual regions
This bug occurred because stretches of low quality (such that s in
line 217 is never above 0) never set beg or end in line 218. This
commit just checks that max has been changed from its default value 0,
and if not (and there's never a high quality base above max's default
value of zero), the beg is 0 and the end 0+max_len.