Add prefixsplit function
This adds the ability to read a fastq/fasta file and split the file based on the prefix of each read to enable fasting sorting of read sets. 

Usage: seqtk prefixsplit [options] <output_filename> <in.fa>
Options:
-p INT    length of prefix 
-A        force FASTA output (discard quality)
-C        drop comments at the header lines

It will create files for each prefix of the specified length, e.g. 
output_filename.AA.fa 
output_filename.AC.fa 
.... 
plus a single file that contains those with an N in the prefix: 
output_filename.N.fa 

There are options to remove the quality scores and drop comments using the same methods as the seqtk seq function. 

I have tried to stick to the coding format of the rest of the file, however, this is my first time coding in C and therefore I am sure there are improvements that could be made.